- Added bsp_epair_t {key,value,next} and bsp_entity_t {epairs,next} to   botlib_structs.h; rewrote AAS_ValueForBSPEpairKey,   AAS_VectorForBSPEpairKey, FloatForKey, AAS_IntForBSPEpairKey,   sub_10006920 (free), sub_100069A0 (parse) and the BotInitLevelItems   entity walk to use them. - Retyped aasworld.d_100669C0 to void** (per-soundindex pointer array);   rewrote InitLevelItemHeap (fixed off-by-one + 4-byte terminator bug),   AllocLevelItem, sub_1002F2B0, AddLevelItemToList, sub_1002F320. - Retyped SetScriptFlags to (script_t*, int) and AAS_DeAllocAASLink   parameters to float*; bulk-fixed all call sites. - Retyped bi_Trace function pointer to typed args. - Fixed sub_100069A0 GetClearedMemory sizes (8 -> 2*ptr, 12 -> 3*ptr).
Source-commit: 1531185c156adf6e76f028f09c08eba023312eca

+++ botlib/botlib_exports.c
@@ -324,16 +324,16 @@ int Export_BotLoadMap(char *mapname, int modelindexes, char **modelindex,
     if (!BotLibSetup("BotLoadMap")) return 1;
 
     if (!mapname) {
-        return BotLoadMap(0, modelindexes, (int)modelindex,
-                          soundindexes, (int)soundindex,
-                          imageindexes, (int)imageindex);
+        return BotLoadMap(0, modelindexes, modelindex,
+                          soundindexes, soundindex,
+                          imageindexes, imageindex);
     }
 
     bi_Print(1, "------------ Map Loading ------------\n");
 
-    errno = BotLoadMap((int)mapname, modelindexes, (int)modelindex,
-                       soundindexes, (int)soundindex,
-                       imageindexes, (int)imageindex);
+    errno = BotLoadMap(mapname, modelindexes, modelindex,
+                       soundindexes, soundindex,
+                       imageindexes, imageindex);
     if (errno) return errno;
 
     sub_10029C10();
